Ill fated attempt to add connection pooling to Paisley.
This commit breaks everything.

Problem is that twisted really wasn't designed with keeping http requests open in mind. My strategy was to try and spoof a connection by opening a tcp connection, and passing that to a HTTPClient manually, but I ran out of time.

Another, potentially better approach would be to reimplement the basc HTTP stuff for couch only.

The best approach would probably be to stop using twisted and condemn it to its rightful place amongst the brain destroying filth of evented python in code hades. Whoever thought that creating an evented framework in a language without proper lambdas, writing that framework in the style of java, and then making everything only accessable with factoryFactoryFactoryFactories was a good idea, deserves to spend the rest of his life wallowing in the worst bugs an asynchronous web framework can encounter. Most code wants to be open, but for the sake of general sanity, Twisted should have been closed source.